backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
multidict==6.4.2
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.19.0
